refactor(domain): 替换配置公开字段为布尔类型
- 将 ConfigDTO 和 ConfigVO 中的 isPublic 字段由 Integer 类型改为 Boolean 类型，命名为 publicFlag
- 更新字段注释，改用 true/false 表示是否对前端公开
- 删除 ydsz-userinfo 模块中的 OAuth2Application 和 SecurityAlert 实体类及相关代码

+++ ydsz-common/ydsz-common-feign/src/main/java/com/njydsz/common/feign/circuitbreaker/FeignCircuitBreakerStrategy.java
@@ -1,5 +1,7 @@
 package com.njydsz.common.feign.circuitbreaker;
 
+import java.math.BigDecimal;
+
 /**
  * Feign 熔断器策略接口。
  *
@@ -31,9 +33,9 @@ public interface FeignCircuitBreakerStrategy {
     /**
      * 获取失败率（百分比）。
      *
-     * @return 失败率（0-100）
+     * @return 失败率（0-100），使用 BigDecimal 避免浮点精度丢失
      */
-    float getFailureRate();
+    BigDecimal getFailureRate();
 
     /**
      * 获取总调用次数。
